finish moving away from zustand to tanstack query/tanstack router
